requirements.txt
# -----------------------------------------------------------------
# Requirements for BLS (Bureau of Labor Statistics) Data Fetcher
# -----------------------------------------------------------------
# Use this file to install the necessary dependencies:
# pip install -r requirements.txt
# -----------------------------------------------------------------

# Core library for making HTTP requests to the BLS API.
requests==2.31.0

# Powerful data analysis and manipulation library, used for creating
# and managing the final data table (DataFrame).
pandas==2.2.2

# Utility for loading environment variables from a .env file,
# used here to securely manage BLS API keys.
python-dotenv==1.0.1